fix: make the clippy gate pass again — main has been red for two commits
`scripts/check.sh` runs `cargo clippy --all-targets -- -D warnings`, and the
last two pushes (cc70205, 0000a55) landed 13 lints that the bare `cargo clippy`
they were presumably checked with only reports as warnings. CI has been failing
on main since 2026-08-22; this restores it. No behaviour changes — every fix is
a mechanical rewrite that preserves the original logic.

anima-assets: three `needless_range_loop` in the new BWT decoder rewritten as
iterator+enumerate, two `unnecessary_cast` on an already-u32 block number, and
one `collapsible_str_replace`.

anima-net: `map_entry` twice — the play-server facet cache uses `Entry::Vacant`
because its fallible open carries a `?` that cannot cross a closure, and the
HasSurfaceOverhead cache uses `entry().or_insert_with()`; `too_many_arguments`
on `has_surface_overhead` gets the same `#[allow]` + reason comment `emit_tiles`
already carries; an `unnecessary_cast` on `MALE_GUMP_OFFSET`, which is already
u32; `if_same_then_else` + `needless_bool` in `multi_component_never_draw`
collapsed to one expression with the truth table preserved; and `drop(look)`
removed — `Look` implements no `Drop`, so with NLL the borrow already ends at
its last use, which the comment there now says instead.

Verified: clippy 0 errors, `scripts/check.sh --skip-desktop` green, and all 619
workspace tests pass.

+++ crates/anima-assets/src/map.rs
@@ -348,7 +348,7 @@ impl MapData {
     fn load_land_block(&mut self, bx: u32, by: u32) -> &Vec<(u16, i8)> {
         let key = (bx << 16) | by;
         if !self.land_cache.contains_key(&key) {
-            let block_num = (bx * self.blocks_y + by) as u32;
+            let block_num = bx * self.blocks_y + by;
             let cells = if let Some(&i) = self.patch_land.get(&block_num) {
                 self.diffs
                     .map_block(i)
@@ -382,7 +382,7 @@ impl MapData {
         if self.statics_cache.contains_key(&key) {
             return &self.statics_cache[&key];
         }
-        let block_num = (bx * self.blocks_y + by) as u32;
+        let block_num = bx * self.blocks_y + by;
         let out = if let Some(&i) = self.patch_sta.get(&block_num) {
             match self.diffs.sta_index(i) {
                 Some((pos, len)) if pos != 0xFFFF_FFFF && len != 0 => {
